requirements.txt
async_lru==1.0.2
pyaes==1.6.1
pyrogram==1.2.9
pysocks==1.7.1
tgcrypto==1.2.2
Pillow>=8.3.2
pytesseract==0.3.8
schedule==1.1.0
